+++ demo/src/main/java/com/cybercode/wubbaduel/controller/UserController.java
@@ -1,0 +1,44 @@
+package com.cybercode.wubbaduel.controller;
+
+import com.cybercode.wubbaduel.model.User;
+import com.cybercode.wubbaduel.service.UserService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
+
+@RestController //Convierte a json automaticamente es magia
+@RequestMapping("api/users")
+public class UserController {
+
+    private final UserService userService;
+
+    @Autowired
+    public UserController(UserService userService) {
+        this.userService = userService;
+    }
+
+    // Crear un nuevo usuario
+    @PostMapping
+    public User createUser(@RequestBody User user) {
+        return userService.createUser(user);
+    }
+
+    // Obtener un usuario por nombre de usuario
+    @GetMapping("/{username}")
+    public User getUserByUsername(@PathVariable String username) {
+        return userService.getUserByUsername(username);
+    }
+
+    // Obtener un usuario por ID
+    @GetMapping("/{id}")
+    public Optional<User> getUserById(@PathVariable Long id) {
+        return userService.getUserById(id);
+    }
+
+    // Eliminar un usuario por ID
+    @DeleteMapping("/{id}")
+    public void deleteUser(@PathVariable Long id) {
+        userService.deleteUser(id);
+    }
+}
