+++ demo/src/main/java/com/cybercode/wubbaduel/model/UserCard.java
@@ -1,4 +1,4 @@
-package models;
+package com.cybercode.wubbaduel.model;
 
 import jakarta.persistence.*;
 import lombok.*;
