+++ demo/src/main/java/com/cybercode/wubbaduel/repositorie/CardRepo.java
@@ -1,0 +1,22 @@
+package com.cybercode.wubbaduel.repositorie;
+
+
+import com.cybercode.wubbaduel.model.Card;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface CardRepo extends JpaRepository<Card, Long> {
+	
+    List<Card> findAll();
+    
+    Card findByName(String name);
+    
+    List<Card> findByCostLessThanEqual(int cost);
+    
+    
+}
