+++ demo/src/main/java/com/cybercode/wubbaduel/repositorie/UserCardRepo.java
@@ -1,7 +1,7 @@
-package repositories;
+package com.cybercode.wubbaduel.repositorie;
 
 
-import models.UserCard;
+import com.cybercode.wubbaduel.model.UserCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import java.util.List;
