+++ demo/src/main/java/com/cybercode/wubbaduel/service/CardService.java
@@ -1,7 +1,7 @@
-package services;
+package com.cybercode.wubbaduel.service;
 
-import models.Card;
-import repositories.CardRepo;
+import com.cybercode.wubbaduel.model.Card;
+import com.cybercode.wubbaduel.repositorie.CardRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
