--- demo/src/main/java/controllers/CharacterCardController.java
@@ -1,5 +1,0 @@
-package controllers;
-
-public class CharacterCardController {
-
-}
