--- demo/src/main/java/controllers/UserCardController.java
@@ -1,5 +1,0 @@
-package controllers;
-
-public class UserCardController {
-
-}
