--- demo/src/main/java/principal/Principal.java
@@ -1,7 +1,0 @@
-package principal;
-
-public class Principal {
-	public static void main(String[] args) {
-        System.out.println("Hellow World");
-    }
-}
